+++ NEWS.md
@@ -5,11 +5,13 @@
   from Posit Package Manager snapshots as being installed from CRAN. (#2104)
 
 * `renv` now only checks for archived packages during installation when
-  `options(renv.install.allowArchivedPackages = TRUE)` is set.
+  `options(renv.install.allowArchivedPackages = TRUE)` is set. This fixes
+  an issue where `renv` could erroneously discover that Rcpp 1.3 was available
+  when using the Posit Package Manager repository.
 
 * Fixed an issue where `renv::dependencies()` could fail with an
   unhelpful error message if the scanned document contained a
-  malformed YAML header. (#2117)
+  YAML header whose contents was not a map. (#2117)
 
 
 # renv 1.1.3